fix/enable class join button when allowed to enter
EnrolledAndTimer component relies on the Timer being the message "joinNow" to become enabled, but without a supplied options argument, timeLeftFormat will return the time remaining up until the start time is reached.  Setting the minutes position of cutoffNumArr and cutoffTextArr to 10 and "joinNow" respectively, means that when all larger units (hours, minutes, and seconds) are 0 and the minutes unit is from 1 to 10, "joinNow" will be the returned message.  Currently, the user will not be able to join the class when there's less than 1 minute or the class is in progress.